Change RPM property type from int to decimal
Updated the `RPM` property in the `ProductionAllotmentDto` class
within the `AvyyanBackend.DTOs.ProAllotDto` namespace from an
`int` to a `decimal`. This change allows for greater precision
in representing RPM values, accommodating fractional values
where necessary.

+++ DTOs/ProAllotDto/ProductionAllotmentDto.cs
@@ -43,7 +43,7 @@ namespace AvyyanBackend.DTOs.ProAllotDto
 		public int MachineId { get; set; }
 		public int NumberOfNeedles { get; set; }
 		public int Feeders { get; set; }
-		public int RPM { get; set; }
+		public decimal RPM { get; set; }
 		public decimal RollPerKg { get; set; }
 		public decimal TotalLoadWeight { get; set; }
 		public decimal TotalRolls { get; set; }
